Moved wrappers and test changes
