--- provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/PractitionerRoleWrapper.java
@@ -1,14 +1,0 @@
-package gov.va.api.health.providerdirectory.service;
-
-import lombok.Builder;
-import lombok.Value;
-
-@Value
-@Builder
-public final class PractitionerRoleWrapper {
-  ProviderResponse providerResponse;
-
-  ProviderContactsResponse providerContactsResponse;
-
-  ProviderSpecialtiesResponse providerSpecialtiesResponse;
-}
