--- provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/PractitionerWrapper.java
@@ -1,12 +1,0 @@
-package gov.va.api.health.providerdirectory.service;
-
-import lombok.Builder;
-import lombok.Value;
-
-@Value
-@Builder
-public final class PractitionerWrapper {
-  ProviderResponse providerResponse;
-
-  ProviderContactsResponse providerContactsResponse;
-}
