+++ provider-directory/src/main/java/gov/va/api/health/providerdirectory/service/controller/practitionerrole/PractitionerRoleWrapper.java
@@ -1,0 +1,17 @@
+package gov.va.api.health.providerdirectory.service.controller.practitionerrole;
+
+import gov.va.api.health.providerdirectory.service.ProviderContactsResponse;
+import gov.va.api.health.providerdirectory.service.ProviderResponse;
+import gov.va.api.health.providerdirectory.service.ProviderSpecialtiesResponse;
+import lombok.Builder;
+import lombok.Value;
+
+@Value
+@Builder
+public final class PractitionerRoleWrapper {
+  ProviderResponse providerResponse;
+
+  ProviderContactsResponse providerContactsResponse;
+
+  ProviderSpecialtiesResponse providerSpecialtiesResponse;
+}
